+++ Assets/Scripts/Game/Gameplay/Models/Weapon/WeaponsCharacteristics.cs
@@ -4,27 +4,27 @@ namespace Game.Gameplay.Models.Weapon
 {
     public class WeaponsCharacteristics
     {
-        private readonly Dictionary<WeaponType, Dictionary<WeaponCharacteristicType, int>> _weaponsCharacteristics = new();
+        private readonly Dictionary<WeaponType, Dictionary<WeaponCharacteristicType, float>> _weaponsCharacteristics = new();
 
-        public int GetCharacteristic(WeaponType weaponType, WeaponCharacteristicType weaponCharacteristicType)
+        public float GetCharacteristic(WeaponType weaponType, WeaponCharacteristicType weaponCharacteristicType)
         {
             var characteristicValues = GetOrCreateCharacteristicValues(weaponType);
 
             return characteristicValues[weaponCharacteristicType];
         }
 
-        public void SetCharacteristic(WeaponType weaponType, WeaponCharacteristicType weaponCharacteristicType, int value)
+        public void SetCharacteristic(WeaponType weaponType, WeaponCharacteristicType weaponCharacteristicType, float value)
         {
             var characteristicValues = GetOrCreateCharacteristicValues(weaponType);
 
             characteristicValues[weaponCharacteristicType] = value;
         }
 
-        private Dictionary<WeaponCharacteristicType, int> GetOrCreateCharacteristicValues(WeaponType weaponType)
+        private Dictionary<WeaponCharacteristicType, float> GetOrCreateCharacteristicValues(WeaponType weaponType)
         {
             if (_weaponsCharacteristics.TryGetValue(weaponType, out var characteristicValues) == false)
             {
-                characteristicValues = new Dictionary<WeaponCharacteristicType, int>();
+                characteristicValues = new Dictionary<WeaponCharacteristicType, float>();
                 _weaponsCharacteristics[weaponType] = characteristicValues;
             }
